requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
matplotlib>=3.6.0
